requirements.txt
pandas
numpy
scipy
matplotlib
seaborn
plotly
python-docx
scikit-learn
statsmodels
openpyxl
xlrd
pyod
networkx
streamlit
